pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["lemon <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
tortoise-orm = "^0.16.14"
aiomysql = "^0.0.20"
aerich = "^0.2.2"
email-validator = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
